make caching layer the standard, remove option to disable it
