+++ doc/Development_Documentation/05_Objects/01_Object_Classes/05_Class_Settings/README.md
@@ -13,4 +13,3 @@ backend or via code by a developer.
 * [Locking Fields](./35_Locking_Fields.md)
 * [Path Formatter](./50_Path_Formatter.md)
 * [Link Generator](./30_Link_Generator.md)
-* [Cache Raw Relation Data](./05_Cache_Raw_Relation_Data.md)
