+++ doc/Development_Documentation/23_Installation_and_Upgrade/09_Upgrade_Notes/README.md
@@ -2,6 +2,9 @@
 
 
 ## 6.5.0
+- Relation Types DB Caching Layer is always turned on now. Removed support for non-cached alternative. 
+  All rows of the affected `object_relation_` table will be fetched in on go and cached together we with the obejct. 
+  see https://github.com/pimcore/pimcore/issues/5427
 - If you have custom lazy-loaded datatypes **not** extending `Pimcore\Model\DataObject\ClassDefinition\Data\Relations\AbstractRelations`, implement the `LazyLoadingSupportInterface`
   The `method_exists('getLazyLoading')` calls will be removed in Pimcore 7.
 - It is now possible to configure `php:memory_limit` for `web2print:pdf-creation` command with following configuration:
