+++ models/DataObject/AbstractObject.php
@@ -300,10 +300,8 @@ class AbstractObject extends Model\Element\AbstractElement
 
                     $object->__setDataVersionTimestamp($object->getModificationDate());
 
-                    if ($object instanceof CacheRawRelationDataInterface) {
-                        // force loading of relation data
-                        $object->__getRawRelationData();
-                    }
+                    // force loading of relation data
+                    $object->__getRawRelationData();
 
                     Cache::save($object, $cacheKey);
                 } else {
