--- models/DataObject/CacheRawRelationDataInterface.php
@@ -1,26 +1,0 @@
-<?php
-/**
- * Pimcore
- *
- * This source file is available under two different licenses:
- * - GNU General Public License version 3 (GPLv3)
- * - Pimcore Enterprise License (PEL)
- * Full copyright and license information is available in
- * LICENSE.md which is distributed with this source code.
- *
- * @category   Pimcore
- * @package    Element
- *
- * @copyright  Copyright (c) Pimcore GmbH (http://www.pimcore.org)
- * @license    http://www.pimcore.org/license     GPLv3 and PEL
- */
-
-namespace Pimcore\Model\DataObject;
-
-interface CacheRawRelationDataInterface
-{
-    /**
-     * @return array
-     */
-    public function __getRawRelationData(): array;
-}
