+++ models/DataObject/ClassDefinition.php
@@ -118,11 +118,6 @@ class ClassDefinition extends Model\AbstractModel
     public $showVariants = false;
 
     /**
-     * @var bool
-     */
-    public $cacheRawRelationData = false;
-
-    /**
      * @var array
      */
     public $fieldDefinitions = [];
@@ -397,21 +392,11 @@ class ClassDefinition extends Model\AbstractModel
         }
         $cd .= "*/\n\n";
 
-        $implementsBlock = '\\Pimcore\\Model\\DataObject\\DirtyIndicatorInterface';
-        if ($this->getCacheRawRelationData()) {
-            $implementsBlock .= ',\\Pimcore\\Model\\DataObject\\CacheRawRelationDataInterface';
-        }
-
-        $cd .= 'class '.ucfirst($this->getName()).' extends '.$extendClass.' implements ' . $implementsBlock . ' {';
+        $cd .= 'class '.ucfirst($this->getName()).' extends '.$extendClass.' implements \\Pimcore\\Model\\DataObject\\DirtyIndicatorInterface {';
         $cd .= "\n\n";
 
         $cd .= 'use \Pimcore\Model\DataObject\Traits\DirtyIndicatorTrait;';
         $cd .= "\n\n";
-
-        if ($this->getCacheRawRelationData()) {
-            $cd .= 'use \Pimcore\Model\DataObject\Traits\CacheRawRelationDataTrait;';
-            $cd .= "\n\n";
-        }
 
         if ($this->getUseTraits()) {
             $cd .= 'use '.$this->getUseTraits().";\n";
@@ -1329,20 +1314,7 @@ class ClassDefinition extends Model\AbstractModel
     /**
      * @return bool
      */
-    public function getCacheRawRelationData(): bool
-    {
-        return $this->cacheRawRelationData;
-    }
-
-    /**
-     * @param bool $cacheRawRelationData
-     *
-     * @return $this
-     */
-    public function setCacheRawRelationData($cacheRawRelationData)
-    {
-        $this->cacheRawRelationData = (bool) $cacheRawRelationData;
-
-        return $this;
+    public function getCacheRawRelationData() {
+        return true;
     }
 }
