+++ tests/_support/Resources/objects/lazyloading/class_LazyLoading_export.json
@@ -8,7 +8,6 @@
     "allowInherit": true,
     "allowVariants": false,
     "showVariants": false,
-    "cacheRawRelationData": false,
     "layoutDefinitions": {
         "fieldtype": "panel",
         "labelWidth": 100,
